UI_agent/requirements.txt
# Core dependencies
streamlit>=1.28.0,<2.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0

# LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Additional utilities
requests>=2.31.0
urllib3>=2.0.0

# Windows-specific alternatives (uncomment if needed)
# faiss-gpu>=1.7.4  # If you have CUDA GPU on Windows
# streamlit==1.32.0  # Specific stable version if latest fails

# Development and testing
# jupyter>=1.0.0  # If running in Jupyter environment
